# sdk/openapi.json
{
  "components": {
    "schemas": {
      "Error": {
        "properties": {
          "error": {
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              }
            },
            "type": "object"
          }
        },
        "type": "object"
      },
      "Event": {
        "properties": {
          "at": {
            "type": "string"
          },
          "from": {
            "type": "string"
          },
          "id": {
            "type": "integer"
          },
          "reason": {
            "type": "string"
          },
          "sandbox_id": {
            "type": "string"
          },
          "to": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "EventList": {
        "items": {
          "$ref": "#/components/schemas/Event"
        },
        "type": "array"
      },
      "ExecResult": {
        "properties": {
          "exit_code": {
            "type": "integer"
          },
          "stderr": {
            "type": "string"
          },
          "stdout": {
            "type": "string"
          },
          "timed_out": {
            "type": "boolean"
          },
          "truncated": {
            "type": "boolean"
          }
        },
        "type": "object"
      },
...
      "Published": {
        "properties": {
          "port": {
            "type": "integer"
          },
          "url": {
            "type": "string"
          },
          "visibility": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "Registry": {
        "properties": {
          "created_at": {
            "type": "string"
          },
          "host": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "RegistryList": {
        "items": {
          "$ref": "#/components/schemas/Registry"
        },
        "type": "array"
      },
      "Sandbox": {
        "properties": {
          "created_at": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "lifecycle": {
            "type": "string"
          },
          "metadata": {
            "type": "object"
          },
          "published": {
            "items": {
              "$ref": "#/components/schemas/Published"
            },
            "type": "array"
          },
          "state": {
            "type": "string"
          },
          "template": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "SandboxList": {
        "items": {
          "$ref": "#/components/schemas/Sandbox"
        },
        "type": "array"
      },
...
    },
    "securitySchemes": {
      "apiKey": {
        "description": "A tenant API key, or an OAuth access token.",
        "scheme": "bearer",
        "type": "http"
      },
      "oauth2": {
        "description": "An OAuth 2.1 access token of this tenant.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Isolated Linux microVMs on one host, created from OCI images and forked into copies.",
    "title": "Kiln",
    "version": "0.3.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/v1/events": {
      "get": {
        "operationId": "events",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventList"
                }
              }
            },
            "description": "The call succeeded."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The request is invalid. The message names the field."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The credential is missing, expired or revoked."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The row does not exist, or it belongs to another tenant."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The call conflicts with the state, for example a template that still has sandboxes."
          },
          "507": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "A cap of this tenant is reached. The message names the cap."
          }
        },
        "summary": "Read the event stream of this tenant."
      }
    },
...
    "/v1/registries": {
      "get": {
        "operationId": "listRegistries",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistryList"
                }
              }
            },
            "description": "The call succeeded."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The request is invalid. The message names the field."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The credential is missing, expired or revoked."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The row does not exist, or it belongs to another tenant."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The call conflicts with the state, for example a template that still has sandboxes."
          },
          "507": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "A cap of this tenant is reached. The message names the cap."
          }
        },
        "summary": "List the registry credentials of this tenant. The tokens are never returned."
      },
      "post": {
        "operationId": "putRegistry",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "host": {
                    "type": "string"
                  },
                  "token": {
                    "type": "string"
                  },
                  "username": {
                    "type": "string"
                  }
                },
                "required": [
                  "host",
                  "username",
                  "token"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Registry"
                }
              }
            },
            "description": "The call succeeded."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The request is invalid. The message names the field."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The credential is missing, expired or revoked."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The row does not exist, or it belongs to another tenant."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The call conflicts with the state, for example a template that still has sandboxes."
          },
          "507": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "A cap of this tenant is reached. The message names the cap."
          }
        },
        "summary": "Store the credential a template build uses to pull a private image for this tenant."
      }
    },
    "/v1/registries/{host}": {
      "delete": {
        "operationId": "deleteRegistry",
        "parameters": [
          {
            "description": "Registry hostname.",
            "in": "path",
            "name": "host",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "The call succeeded."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The request is invalid. The message names the field."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The credential is missing, expired or revoked."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The row does not exist, or it belongs to another tenant."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The call conflicts with the state, for example a template that still has sandboxes."
          },
          "507": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "A cap of this tenant is reached. The message names the cap."
          }
        },
        "summary": "Remove one registry credential."
      }
    },
    "/v1/sandboxes": {
      "get": {
        "operationId": "listSandboxes",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SandboxList"
                }
              }
            },
            "description": "The call succeeded."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The request is invalid. The message names the field."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The credential is missing, expired or revoked."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The row does not exist, or it belongs to another tenant."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The call conflicts with the state, for example a template that still has sandboxes."
          },
          "507": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "A cap of this tenant is reached. The message names the cap."
          }
        },
        "summary": "List the sandboxes of this tenant."
      },
      "post": {
        "operationId": "createSandbox",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "idle_seconds": {
                    "type": "integer"
                  },
                  "lifecycle": {
                    "type": "string"
                  },
                  "metadata": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "secrets": {
                    "items": {
                      "type": "string"
                    },
                    "type": "array"
                  },
                  "template": {
                    "type": "string"
                  },
                  "ttl_seconds": {
                    "type": "integer"
                  }
                },
                "required": [
                  "template"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Sandbox"
                }
              }
            },
            "description": "The call succeeded."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The request is invalid. The message names the field."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The credential is missing, expired or revoked."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The row does not exist, or it belongs to another tenant."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The call conflicts with the state, for example a template that still has sandboxes."
          },
          "507": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "A cap of this tenant is reached. The message names the cap."
          }
        },
        "summary": "Create a sandbox by restoring a template snapshot."
      }
    },
    "/v1/sandboxes/{id}": {
      "delete": {
        "operationId": "deleteSandbox",
        "parameters": [
          {
            "description": "Sandbox id.",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "The call succeeded."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The request is invalid. The message names the field."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The credential is missing, expired or revoked."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The row does not exist, or it belongs to another tenant."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The call conflicts with the state, for example a template that still has sandboxes."
          },
          "507": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "A cap of this tenant is reached. The message names the cap."
          }
        },
        "summary": "Destroy a sandbox and retire its hostnames."
      },
      "get": {
        "operationId": "getSandbox",
        "parameters": [
          {
            "description": "Sandbox id.",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Sandbox"
                }
              }
            },
            "description": "The call succeeded."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The request is invalid. The message names the field."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The credential is missing, expired or revoked."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The row does not exist, or it belongs to another tenant."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The call conflicts with the state, for example a template that still has sandboxes."
          },
          "507": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "A cap of this tenant is reached. The message names the cap."
          }
        },
        "summary": "Read one sandbox with its published ports."
      }
    },
    "/v1/sandboxes/{id}/exec": {
      "post": {
        "operationId": "exec",
        "parameters": [
          {
            "description": "Sandbox id.",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "cmd": {
                    "items": {
                      "type": "string"
                    },
                    "type": "array"
                  },
                  "cwd": {
                    "type": "string"
                  },
                  "env": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "timeout_seconds": {
                    "type": "integer"
                  }
                },
                "required": [
                  "cmd"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecResult"
                }
              }
            },
            "description": "The call succeeded."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The request is invalid. The message names the field."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The credential is missing, expired or revoked."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The row does not exist, or it belongs to another tenant."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The call conflicts with the state, for example a template that still has sandboxes."
          },
          "507": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "A cap of this tenant is reached. The message names the cap."
          }
        },
        "summary": "Run a command inside a sandbox. Accept text/event-stream to read the output as it runs."
      }
    },
...
    "/v1/sandboxes/{id}/fork": {
      "post": {
        "operationId": "fork",
        "parameters": [
          {
            "description": "Sandbox id.",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "allow_secret_fork": {
                    "type": "boolean"
                  },
                  "count": {
                    "type": "integer"
                  }
                },
                "required": [
                  "count"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SandboxList"
                }
              }
            },
            "description": "The call succeeded."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The request is invalid. The message names the field."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The credential is missing, expired or revoked."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The row does not exist, or it belongs to another tenant."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The call conflicts with the state, for example a template that still has sandboxes."
          },
          "507": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "A cap of this tenant is reached. The message names the cap."
          }
        },
        "summary": "Snapshot a running sandbox and restore several copies of it."
      }
    },
...
  },
  "security": [
    {
      "apiKey": []
    },
    {
      "oauth2": []
    }
  ],
  "servers": [
    {
      "url": "https://api.example.com"
    }
  ]
}